+++ app/Enums/RegisterTypeEnum.php
@@ -1,0 +1,9 @@
+<?php
+
+namespace App\Enums;
+
+enum RegisterTypeEnum: string
+{
+    case CANCEL = 'cancel';
+    case AUTHORIZE = 'authorize';
+}
