+++ app/Http/Controllers/Api/HealthController.php
@@ -7,12 +7,15 @@ use Illuminate\Http\JsonResponse;
 
 class HealthController extends Controller
 {
+    /**
+     * Check application health.
+     * 
+     * @return JsonResponse
+     */
     public function check(): JsonResponse
     {
         return response()->json(
-            [
-            'status' => 200,
-            ]
+            ['status' => 200]
         );
     }
 }
