+++ config/services.php
@@ -30,4 +30,12 @@ return [
         'region' => env('AWS_DEFAULT_REGION', 'us-east-1'),
     ],
 
+    'authorizing-service' => [
+        'base_uri' => env('URL_AUTHORIZING_SERVICE', 'https://run.mocky.io/v3/8fafdd68-a090-496f-8c9a-3442cf30dae6'),
+    ],
+
+    'send-email-service' => [
+        'base_uri' => env('URL_SEND_EMAIL', 'http://o4d9z.mocklab.io/notify'),
+    ],
+
 ];
